Delete the old set-tracker and fix up tests
